requirements.txt
Flask>=2.2.2
Pillow>=9.0.0